refactor(pr-lifecycle): absorb pr-description.ts into pr-lifecycle.ts
Move all PR description logic (commit categorization, Closes block
building, PR body builders) from pr-description.ts into pr-lifecycle.ts,
giving the PR lifecycle module a complete, self-contained implementation.

- Inline all types and functions from pr-description.ts
- Delete pr-description.ts and its three test files
- Merge pr-description tests into pr-lifecycle.test.ts
- Rename learnings/review-pass tests to pr-lifecycle-* prefix
- Update pr-closes.test.ts and issues.ts imports/comments
- Update ARCHITECTURE.md to reflect consolidated module

Closes #430
